[ADD] makeRactangle_Action & UIViewExtention

+++ DrawingApp/DrawingApp/ShapeElement/RGB.swift
@@ -13,20 +13,20 @@ final class RGB:CustomStringConvertible{
     }
     
     //min과 max는 타입 자체와 관련이 있기 때문에 Static으로 선언했습니다.
-    static let minValue: Int = 0
-    static let maxValue: Int = 255
+    static let minValue: Double = 0
+    static let maxValue: Double = 255
     
-    private var inputRedValue:Int
-    private var inputGreenValue:Int
-    private var inputBlueValue:Int
+    private var inputRedValue:Double
+    private var inputGreenValue:Double
+    private var inputBlueValue:Double
     
 
-    var red:Int { self.translateInputRGBValue(inputValue: inputRedValue) }
-    var green:Int { self.translateInputRGBValue(inputValue: inputGreenValue) }
-    var blue:Int { self.translateInputRGBValue(inputValue: inputBlueValue) }
+    var red:Double { self.translateInputRGBValue(inputValue: inputRedValue) }
+    var green:Double { self.translateInputRGBValue(inputValue: inputGreenValue) }
+    var blue:Double { self.translateInputRGBValue(inputValue: inputBlueValue) }
     
     
-    init(red:Int,green:Int,blue:Int) {
+    init(red:Double,green:Double,blue:Double) {
         self.inputRedValue = red
         self.inputGreenValue = green
         self.inputBlueValue = blue
@@ -34,7 +34,7 @@ final class RGB:CustomStringConvertible{
     
     
     //가장큰값 maxValue보다 크면 maxValue로 가작 작은값 minValue보다 작으면 minValue를 주도록 했습니다.
-    private func translateInputRGBValue(inputValue:Int) -> Int{
+    private func translateInputRGBValue(inputValue:Double) -> Double{
         
         if inputValue < RGB.minValue {
             return RGB.minValue
